apps -> order -> submit form added "final"

+++ apps/order/views.py
@@ -69,6 +69,7 @@ class SubmitOrderService(FormView):
         return kwargs
 
     def form_valid(self, form):
+        final = form.cleaned_data['final']
         order_id = form.cleaned_data['order_id']
         service = Service.objects.get(id=self.kwargs['pk'])
         if not self.request.user.is_superuser:
@@ -88,30 +89,34 @@ class SubmitOrderService(FormView):
             order = Order(customer=customer, service=service)
             order.save()
             content_data = []
-        try:
-            if order_id != -1:
-                with open(order.file.path, 'w+') as f:
-                    f.truncate()
-                    writer = csv.writer(f)
-                    for i in range(len(content_data)):
-                        if i != order_id-1:
-                            writer.writerow(content_data[i])
-            else:
-                fields = csv.reader(open(service.fields.path, 'r'))
-                data = []
-                for row in fields:
-                    data.append(form.cleaned_data[row[0]])
-                content_data.append(data)
+        if str(final) == "1":
+            order.is_finished = True
+            self.success_url = reverse_lazy("index:index")
+        else:
+            try:
+                if order_id != -1:
+                    with open(order.file.path, 'w+') as f:
+                        f.truncate()
+                        writer = csv.writer(f)
+                        for i in range(len(content_data)):
+                            if i != order_id-1:
+                                writer.writerow(content_data[i])
+                else:
+                    fields = csv.reader(open(service.fields.path, 'r'))
+                    data = []
+                    for row in fields:
+                        data.append(form.cleaned_data[row[0]])
+                    content_data.append(data)
 
-                with open(order.file.path, 'a') as f:
+                    with open(order.file.path, 'a') as f:
+                        writer = csv.writer(f)
+                        writer.writerow(data)
+            except:
+                with open(order.file.path, 'w') as f:
                     writer = csv.writer(f)
-                    writer.writerow(data)
-        except:
-            with open(order.file.path, 'w') as f:
-                writer = csv.writer(f)
-                for row in content_data:
-                    writer.writerow(row)
-        self.success_url = reverse_lazy('order:order_service', kwargs={'pk': self.kwargs['pk']})
+                    for row in content_data:
+                        writer.writerow(row)
+            self.success_url = reverse_lazy('order:order_service', kwargs={'pk': self.kwargs['pk']})
         return super(SubmitOrderService, self).form_valid(form)
 
     def form_invalid(self, form):
